Hello Heating and Cooling Forecast
First successful run at generating plots for heating and cooling degree
day forecasts.

R package which supports:

* Loading data from the National Weather Forecast Database
* Calculating heating and cooling degree hours for each forecast
interval, using a naive (but somewhat standard) balance point
algorithm.
* Generating plots for heating and cooling degree hours for each
interval, in addition to aggregate animated GIF and MPEG plots.
* Experimenting with sample data, taken on March 19, 2014 and stored
along with the package.

Caveats:

* I can't seem to be able to get the start time interval from the NOAA
forecast. I see it plain as day in the GRIB viewer, but `rgdal` doesn't
seem to be loading it. Need to fix this, or at least work around it
by creating an end-to-end flow which retrieves the start time.
* I'm not doing anything in terms of plot legends. Next, I'd like to add
a legend and posibly some contour values.
* The naive heating and cooling degree hours are alright, but they'r
difficult to understand visually as-is. I need to add better context as
to what people are looking at.