Hello 7-day forecast
Generalize the previous functionality to work with forecasts for days
1 to 3 and 4 to 7. This involves support for different forecast
durations, as well as different NDFD end points.

Add a function for loading all data and generating all plots. I ran this
once and introduced all of the plots to this repo, for the sake of a
demo.

Also: fix a few bugs with file naming, add minimal validation, and improve
some documentation.